refactor: move obtain to examples module
